Initial refactor of querying model into a separate project module
