+++ server/exercise-query/lib/CVC4.jar
